dom/base/test/test_popup_middle_click.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for triggering popup by middle-click mouse events</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<div id="target" style="width: 50px; height: 50px; background: green"></div>
<script>

function sendMouseEvent(element, eventName, listenEventName, handler) {
  element.addEventListener(listenEventName, handler, {once: true});
  synthesizeMouseAtCenter(element, {type: eventName, button: 0});

  element.addEventListener(listenEventName, handler, {once: true});
  synthesizeMouseAtCenter(element, {type: eventName, button: 1});
}

function checkAllowOpenPopup(e) {
  let w = window.open("about:blank");
  ok(w, "Should allow popup in the " + e.type + " listener");
  if (w) {
    w.close();
  }
}

function checkBlockOpenPopup(e) {
  let w = window.open("about:blank");
  ok(!w, "Should block popup in the " + e.type + " listener");
  if (w) {
    w.close();
  }
}

function startTest() {
  let target = document.getElementById("target");
  sendMouseEvent(target, "mousemove", "mousemove", checkBlockOpenPopup);
  sendMouseEvent(target, "mousedown", "mousedown", checkBlockOpenPopup);
  sendMouseEvent(target, "mousemove", "mousemove", checkBlockOpenPopup);
  sendMouseEvent(target, "mouseup", "mouseup", checkAllowOpenPopup);
}

const DENY_ACTION = SpecialPowers.Ci.nsIPermissionManager.DENY_ACTION;

let xorigin = SimpleTest.getTestFileURL("").replace(location.hostname, 'mochi.xorigin-test');
add_task(async () => {
  await SpecialPowers.pushPermissions([
    {'type': 'popup', 'allow': DENY_ACTION,
     'context': document},
    {'type': 'popup', 'allow': DENY_ACTION,
     'context': xorigin}
  ]);

  await new Promise(resolve => SimpleTest.waitForFocus(resolve));

  startTest();
});

</script>
</body>
</html>
